Refactor profiles theme GH-224
