+++ .changeset/gorgeous-wombats-dance.md
@@ -1,0 +1,5 @@
+---
+'@undataforum/gatsby-theme-profiles': minor
+---
+
+Refactor theme-profiles to enable new features
