+++ themes/gatsby-theme-profiles/index.js
@@ -1,2 +1,2 @@
-export { default as ProfilesPage } from './src/components/profiles-page';
-export { normalize } from './src/templates/profile';
+// ProfilesPage is required for special profile overview pages such as the committee pages.
+export { default as ProfilesPage } from './src/components/profiles';
