+++ themes/gatsby-theme-profiles/package.json
@@ -31,12 +31,13 @@
     "theme-ui": "^0.3.1"
   },
   "dependencies": {
-    "@undataforum/components": "^0.70.2",
+    "@undataforum/components": "^0.71.0",
     "@undataforum/gatsby-theme-base": "^0.12.1",
-    "gatsby-image": "^2.2.38",
-    "gatsby-plugin-sharp": "^2.4.4",
-    "gatsby-source-filesystem": "^2.1.47",
-    "gatsby-transformer-sharp": "^2.3.12",
+    "gatsby-core-utils": "^1.0.28",
+    "gatsby-image": "^2.2.41",
+    "gatsby-plugin-sharp": "^2.4.5",
+    "gatsby-source-filesystem": "^2.1.48",
+    "gatsby-transformer-sharp": "^2.3.14",
     "prop-types": "^15.7.2",
     "remark": "^11.0.1",
     "strip-markdown": "^3.1.0"
